[WIN32] fixed access denied error when trying to create the directory
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@30938 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ project/Win32BuildSetup/BuildSetup.bat
@@ -147,7 +147,7 @@ IF %comp%==vs2008 (
 
 :MAKE_BUILD_EXE
   ECHO Copying files...
-  rmdir BUILD_WIN32 /S /Q
+  IF EXIST BUILD_WIN32 rmdir BUILD_WIN32 /S /Q
   md BUILD_WIN32\Xbmc
 
   Echo .svn>exclude.txt
